copy over base package state for debug one
Fixes a small regression in 0cec05b ("build initramfs: download
packages into per-flavor dirs"), which while taking a better
approach in general, missed that the debug package list is selected
with the assumption that it builds on top of the base list, but with
the split into two separate directories (to fix pre-caching), this
was lost.

While the packages selected pulled in a surprising amount of reverse
dependencies, it was still missing some like lvm2 tools (dmeventd and
co) and most notably busybox - which is really nice to have in the
debug image.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>